Assistente Nexforce/requirements.txt
pandas
pyttsx3
openai
ffmpeg-python
PyAudio
whisper-openai
SpeechRecognition